moved around swagger file
