+++ src/app/api/docs/route.ts
@@ -1,0 +1,7 @@
+// src/app/api/docs/route.ts
+import { getApiDocs } from '@/lib/swagger';
+import { NextResponse } from 'next/server';
+
+export async function GET() {
+  return NextResponse.json(getApiDocs());
+}
